requirements.txt
tflite-runtime
numpy<2.0.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
